requirements.txt
sounddevice==0.5.1
numpy==2.0.2
scipy==1.14.1
python-dotenv==1.0.1
numba==0.60.0
librosa==0.10.2.post1 
colorama==0.4.6
elevenlabs==1.12.1
keyboard==0.13.5
